Move extensions around;
