--- src/Aer.QdrantClient.Http/Helpers/NetstandardPolyfill/HttpRequestExceptionExtensions.cs
@@ -1,22 +1,0 @@
-﻿using System.Net;
-
-namespace Aer.QdrantClient.Http.Helpers.NetstandardPolyfill;
-
-#if NETSTANDARD2_0
-internal static class HttpRequestExceptionExtensions
-{
-	private const string StatusCodeKeyName = "StatusCode";
-
-	public static bool SetStatusCode(this HttpRequestException httpRequestException, HttpStatusCode httpStatusCode)
-	{
-		httpRequestException.Data[StatusCodeKeyName] = httpStatusCode;
-
-		return false;
-	}
-
-	public static HttpStatusCode? GetStatusCode(this HttpRequestException httpRequestException)
-	{
-		return (HttpStatusCode?) httpRequestException.Data[StatusCodeKeyName];
-	}
-}
-#endif
