+++ src/Aer.QdrantClient.Http/Helpers/NetstandardPolyfill/Netstandard20Extensions.cs
@@ -1,4 +1,6 @@
-﻿namespace Aer.QdrantClient.Http.Helpers.NetstandardPolyfill;
+﻿using System.Net;
+
+namespace Aer.QdrantClient.Http.Helpers.NetstandardPolyfill;
 
 #if NETSTANDARD2_0
 internal static class Netstandard20Extensions
@@ -28,6 +30,37 @@ internal static class Netstandard20Extensions
 	public static bool Contains(this string target, string value, StringComparison comparisonType)
 	{
 		return target.IndexOf(value, comparisonType) >= 0;
+	}
+}
+
+internal static class HttpResponseMessageExtensions
+{
+	private const string StatusCodeKeyName = "StatusCode";
+
+	public static bool SetStatusCode(this HttpRequestException httpRequestException, HttpStatusCode httpStatusCode)
+	{
+		httpRequestException.Data[StatusCodeKeyName] = httpStatusCode;
+
+		return false;
+	}
+
+	public static HttpStatusCode? GetStatusCode(this HttpRequestException httpRequestException)
+	{
+		return (HttpStatusCode?) httpRequestException.Data[StatusCodeKeyName];
+	}
+
+	public static HttpResponseMessage SetStatusCode(this HttpResponseMessage httpResponseMessage)
+	{
+		try
+		{
+			httpResponseMessage.EnsureSuccessStatusCode();
+		}
+		catch (HttpRequestException ex) when (ex.SetStatusCode(httpResponseMessage.StatusCode))
+		{
+			// Intentionally left empty. Will never be reached.
+		}
+
+		return httpResponseMessage;
 	}
 }
 #endif
